requirements.txt
pyrogram==1.4.16
py-tgcalls==0.8.6 
TgCrypto
git+https://github.com/Sorrybabe/downloader
python-dotenv
youtube_dl
youtube_search_python 
requests
aiohttp
aiofiles
asyncio
youtube_search
python-arq
search_engine_parser
wget 
ffmpeg-python
Pillow
ujson
pymongo[srv]
psutil
motor
